Laravel From Scratch [Part 11]

+++ app/Http/Controllers/PostsController.php
@@ -9,6 +9,16 @@ use App\Post;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -38,15 +48,17 @@ class PostsController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
-          'title' => 'required',
-          'body' => 'required'
-        ]);
+        $this->validate(
+            $request, [
+            'title' => 'required',
+            'body' => 'required'
+            ]
+        );
 
         // Create Post
         $post = new Post;
@@ -60,7 +72,7 @@ class PostsController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
@@ -72,31 +84,44 @@ class PostsController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // Check for the right user_id
+        if (auth()->user()->id != $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  \Illuminate\Http\Request $request
+     * @param  int                      $id
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id)
     {
-        $this->validate($request, [
-          'title' => 'required',
-          'body' => 'required'
-        ]);
+        $this->validate(
+            $request, [
+            'title' => 'required',
+            'body' => 'required'
+            ]
+        );
 
         // Update Post
         $post = Post::find($id);
+
+        // Check for the right user_id
+        if (auth()->user()->id != $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
+        
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->save();
@@ -107,12 +132,16 @@ class PostsController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
         $post = Post::find($id);
+        // Check for the right user_id
+        if (auth()->user()->id != $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post Deleted');
     }
